pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "gendiff"
authors = ["qamaz <solgrini@yandex.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
